slightly re-organize tests by moving them to a proper package
